+++ src/App.js
@@ -1,7 +1,5 @@
 import { RouterProvider } from "react-router-dom";
 import { routers } from "./router";
-import { LocationObserver } from "./utils/hooks/store/useLocationObserver";
-
 /**
  * App 컴포넌트는 애플리케이션의 루트 컴포넌트입니다.
  * RouterProvider를 사용하여 라우터 설정을 적용합니다.
@@ -9,11 +7,7 @@ import { LocationObserver } from "./utils/hooks/store/useLocationObserver";
  * @returns JSX.Element - 라우터 설정이 적용된 App 컴포넌트
  */
 function App() {
-  return (
-    <RouterProvider router={routers}>
-      <LocationObserver />
-    </RouterProvider>
-  );
+  return <RouterProvider router={routers} />;
 }
 
 export default App;
