+++ src/utils/hooks/store/useAuthStore.js
@@ -11,11 +11,10 @@ const useAuthStore = create(
     accessToken: null,
     isAuthenticated: false,
     id: null,
-    username: null,
+    userName: null,
     setAccessToken: (token) =>
-      set({ accessToken: token, isAuthenticated: !!token }),
-    setUser: ({ id, username, token }) =>
-      set({ id, username, accessToken: token, isAuthenticated: !!token }),
+      set({ accessToken: token, isAuthenticated: true }),
+    setUser: ({ id, userName }) => set({ id, userName }),
     logout: async () => {
       try {
         // 서버의 로그아웃 엔드포인트에 요청
@@ -24,7 +23,12 @@ const useAuthStore = create(
         console.error("Logout failed:", error);
       }
       // 클라이언트 상태 업데이트
-      set({ accessToken: null, isAuthenticated: false });
+      set({
+        accessToken: null,
+        isAuthenticated: false,
+        userName: null,
+        id: null
+      });
     }
   }))
 );
